[BUGFIX] Cast floor’ed division to int
The division is rounded by a `floor()` call. Since `floor()` returns a
value of type `double`, the result must get casted to an integer to have
the strict comparisons functional again.

Resolves: #84782
Related: #83556
Releases: master
Reviewed-on: https://review.typo3.org/56726
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Frank Naegler <frank.naegler@typo3.org>

+++ typo3/sysext/backend/Classes/Form/Element/AbstractFormElement.php
@@ -341,7 +341,7 @@ abstract class AbstractFormElement extends AbstractNode
      */
     protected function calculateColumnMarkup(int $cols): array
     {
-        $colWidth = floor(12 / $cols);
+        $colWidth = (int)floor(12 / $cols);
         $colClass = 'col-md-12';
         $colClear = [];
         if ($colWidth === 6) {
